+++ src/senaite/core/catalog/sample_catalog.py
@@ -79,7 +79,6 @@ COLUMNS = BASE_COLUMNS + [
     "getDescendantsUIDs",
     "getDistrict",
     "getDueDate",
-    "getHazardous",
     "getInternalUse",
     "getInvoiceExclude",
     "getPhysicalPath",